First major rewrite of administrative pages.
Both in German and English websites, we now have:

* A new homepage (index page) with a modest introduction
* No more news page; the blog takes it position for now
* A team page. Will go again if no others volunteer to appear there
* No more "things we look for" page because doesn't apply for the moment
* No more projects, teaching, etc. because outdated

These old topics (didactics, robotics, etc.) moved to the side navigation
for the time being, with big warning boxes. That also means that
old pages such as robotics, lehrerinfo, got reactivated.

Furthermore, DE and EN are now again in sync.

This edit was drafted for a couple of weeks. For the content
itself, it is not a major thing, but a number of gimmicks were
added, such as automatic blog post listing (useless for one
post, but more are in the pipe) as well as a central team
description in each language as XML file.