Adding Docker support.
Minimal Dockerfile to run the website out of a LAMP container. This allows to run it on any Linux/Mac OS X system with zero extra installation and totally controllable (when the docker container is stopped, there are no daemons left running on the developer machine). Should also work on modern Windows (shouldn't it?).